// examples-al/workflow/server-blended-modeling/src/language-server/generated/ast.ts
export type MetaInfo = Position | Size;

export const MetaInfo = 'MetaInfo';

export function isMetaInfo(item: unknown): item is MetaInfo {
    return reflection.isInstance(item, MetaInfo);
}

export type Node = ActivityNode | Category | TaskNode;

export const Node = 'Node';

export function isNode(item: unknown): item is Node {
    return reflection.isInstance(item, Node);
}

export type NodeType = 'decision' | 'fork' | 'join' | 'merge';

export function isNodeType(item: unknown): item is NodeType {
    return item === 'decision' || item === 'fork' || item === 'join' || item === 'merge';
}

export type TaskType = 'automated' | 'manual';

export function isTaskType(item: unknown): item is TaskType {
    return item === 'automated' || item === 'manual';
}

export type Weight = 'high' | 'low' | 'medium';

export function isWeight(item: unknown): item is Weight {
    return item === 'low' || item === 'medium' || item === 'high';
}

export interface ActivityNode extends AstNode {
    readonly $container: Model;
    readonly $type: 'ActivityNode';
    name: string
    nodeType?: NodeType
}

export const ActivityNode = 'ActivityNode';

export function isActivityNode(item: unknown): item is ActivityNode {
    return reflection.isInstance(item, ActivityNode);
}

export interface Category extends AstNode {
    readonly $container: Model;
    readonly $type: 'Category';
    children?: Model
    label?: string
    name: string
}

export const Category = 'Category';

export function isCategory(item: unknown): item is Category {
    return reflection.isInstance(item, Category);
}

export interface Edge extends AstNode {
    readonly $type: 'Edge' | 'WeightedEdge';
    source: Reference<Node>
    target?: Reference<Node>
}

export const Edge = 'Edge';

export function isEdge(item: unknown): item is Edge {
    return reflection.isInstance(item, Edge);
}

export interface Model extends AstNode {
    readonly $container: Category;
    readonly $type: 'Model';
    edges: Array<Edge>
    metaInfos: Array<MetaInfo>
    nodes: Array<Node>
}

export const Model = 'Model';

export function isModel(item: unknown): item is Model {
    return reflection.isInstance(item, Model);
}

export interface Position extends AstNode {
    readonly $container: Model;
    readonly $type: 'Position';
    node: Reference<Node>
    x: number
    y: number
}

export const Position = 'Position';

export function isPosition(item: unknown): item is Position {
    return reflection.isInstance(item, Position);
}

export interface Size extends AstNode {
    readonly $container: Model;
    readonly $type: 'Size';
    height: number
    node: Reference<Node>
    width: number
}

export const Size = 'Size';

export function isSize(item: unknown): item is Size {
    return reflection.isInstance(item, Size);
}

export interface TaskNode extends AstNode {
    readonly $container: Model;
    readonly $type: 'TaskNode';
    duration?: number
    expanded: boolean
    label?: string
    name: string
    taskType?: TaskType
}

export const TaskNode = 'TaskNode';